+++ wbs/msvc/Basic.vcxproj
@@ -184,6 +184,7 @@
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="..\src\Basic\AdvancedNormalStation.cpp" />
     <ClCompile Include="..\src\Basic\AmountPpt.cpp" />
     <ClCompile Include="..\src\Basic\ANN\ANN.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">NotUsing</PrecompiledHeader>
@@ -336,6 +337,7 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="..\src\Basic\AdvancedNormalStation.h" />
     <ClInclude Include="..\src\Basic\AmountPpt.h" />
     <ClInclude Include="..\src\Basic\ANN\ANN.h" />
     <ClInclude Include="..\src\Basic\ANN\ANNperf.h" />
