+++ wbs/src/UI/WGInputAnalysisDlg.cpp
@@ -35,6 +35,8 @@ namespace WBSF
 	BEGIN_MESSAGE_MAP(CWGInputAnalysisDlg, CDialog)
 		ON_CBN_SELCHANGE(IDC_KIND, &UpdateCtrl)
 		ON_BN_CLICKED(IDC_EXPORT_MATCH, &UpdateCtrl)
+		ON_BN_CLICKED(IDC_EXPORT_NORMALS, &UpdateCtrl)
+		
 	END_MESSAGE_MAP()
 
 
@@ -55,9 +57,12 @@ namespace WBSF
 		DDX_Text(pDX, IDC_INTERNAL_NAME, m_analysis.m_internalName);
 		DDX_CBIndex(pDX, IDC_KIND, m_analysis.m_kind);
 		DDX_Check(pDX, IDC_EXPORT_MATCH, m_analysis.m_bExportMatch);
+		DDX_Check(pDX, IDC_EXPORT_NORMALS, m_analysis.m_bExportNormal);
 		DDX_Text(pDX, IDC_MATCH_NAME, m_analysis.m_matchName);
+		DDX_Text(pDX, IDC_DB_NAME, m_analysis.m_normalsName);
 		DDX_Control(pDX, IDC_KIND, m_kindCtrl);
 		DDX_Control(pDX, IDC_EXPORT_MATCH, m_exportCtrl);
+		DDX_Control(pDX, IDC_EXPORT_NORMALS, m_exportNormalCtrl);
 
 		UpdateCtrl();
 	}
@@ -66,13 +71,22 @@ namespace WBSF
 
 	void CWGInputAnalysisDlg::UpdateCtrl()
 	{
-		bool bEnable = m_kindCtrl.GetCurSel()== CWGInputAnalysis::MATCH_STATION_NORMALS|| m_kindCtrl.GetCurSel()== CWGInputAnalysis::MATCH_STATION_OBSERVATIONS;
+		bool bShowExportNormal = m_kindCtrl.GetCurSel() == CWGInputAnalysis::EXTRACT_NORMALS;
+		bool bEnable = m_kindCtrl.GetCurSel() == CWGInputAnalysis::MATCH_STATION_NORMALS || m_kindCtrl.GetCurSel() == CWGInputAnalysis::MATCH_STATION_OBSERVATIONS;
 		GetDlgItem(IDC_CMN_STATIC1)->EnableWindow(bEnable);
 		GetDlgItem(IDC_EXPORT_MATCH)->EnableWindow(bEnable);
 
-		bool bEnableMatch = bEnable && m_exportCtrl.GetCheck()!=0;
+		GetDlgItem(IDC_EXPORT_MATCH)->ShowWindow(bShowExportNormal ? SW_HIDE : SW_SHOW);
+		GetDlgItem(IDC_MATCH_NAME)->ShowWindow(bShowExportNormal ? SW_HIDE : SW_SHOW);
+		GetDlgItem(IDC_EXPORT_NORMALS)->ShowWindow(bShowExportNormal ? SW_SHOW: SW_HIDE);
+		GetDlgItem(IDC_DB_NAME)->ShowWindow(bShowExportNormal ? SW_SHOW : SW_HIDE);
+
+
+		bool bEnableMatch = bEnable && m_exportCtrl.GetCheck() != 0 || m_exportNormalCtrl.GetCheck() != 0;
 		GetDlgItem(IDC_CMN_STATIC2)->EnableWindow(bEnableMatch);
 		GetDlgItem(IDC_MATCH_NAME)->EnableWindow(bEnableMatch);
+
+		
 
 	}
