+++ wbs/src/WeatherBasedSimulationUI.rc
@@ -393,7 +393,7 @@ BEGIN
                     "Button",BS_AUTOCHECKBOX | BS_MULTILINE | WS_TABSTOP,226,233,128,14
     COMBOBOX        IDC_FIT_OUTPUT_AS_INPUT,235,249,197,158,CBS_DROPDOWN | WS_VSCROLL | WS_TABSTOP
     LTEXT           "Nom du stade adulte",IDC_STATIC,226,321,66,8
-    EDITTEXT        IDC_FIT_ADULT_NAME,302,319,40,12,ES_AUTOHSCROLL
+    EDITTEXT        IDC_FIT_ADULT_NAME,302,319,130,12,ES_AUTOHSCROLL
     LTEXT           "Tmin",IDC_STATIC,308,270,16,8
     LTEXT           "Tmax",IDC_STATIC,355,270,18,8
     CONTROL         "Contraindre T (lo)",IDC_FIT_LIMIT_TB,"Button",BS_AUTOCHECKBOX | BS_MULTILINE | WS_TABSTOP,226,280,70,14
@@ -417,7 +417,7 @@ BEGIN
     CONTROL         "Éviter un taux nul dans l'intervalle des températures observées",IDC_FIT_AVOID_INF_TIME_IN_OBS_RANGE,
                     "Button",BS_AUTOCHECKBOX | BS_MULTILINE | WS_TABSTOP,226,363,208,14
     CONTROL         "Utiliser les morts",IDC_FIT_USE_DEAD,"Button",BS_AUTOCHECKBOX | BS_MULTILINE | WS_TABSTOP,357,380,75,14
-    LTEXT           "Les adultes n'ont pas de T (hi)",IDC_STATIC,348,318,60,20
+    LTEXT           "Les adultes n'ont pas de T (hi)",IDC_STATIC,235,332,155,12
     COMBOBOX        IDC_FIT_SA_PRESET,285,178,119,158,CBS_DROPDOWN | WS_VSCROLL | WS_TABSTOP
     COMBOBOX        IDC_FIT_METHOD,308,196,124,158,CBS_DROPDOWN | WS_VSCROLL | WS_TABSTOP
     LTEXT           "Précision/temps",IDC_STATIC,226,179,54,9
@@ -1658,11 +1658,13 @@ BEGIN
     GROUPBOX        "",IDC_CMN_STATIC1,7,68,355,40
     CONTROL         "Exporte stations appariées",IDC_EXPORT_MATCH,"Button",BS_AUTOCHECKBOX | WS_TABSTOP,13,68,99,10
     LTEXT           "Nom :",IDC_CMN_STATIC2,18,88,20,8
-    EDITTEXT        IDC_MATCH_NAME,53,85,296,14,ES_AUTOHSCROLL
+    EDITTEXT        IDC_MATCH_NAME,53,84,296,14,ES_AUTOHSCROLL
     LTEXT           "Nom interne :",IDC_STATIC,7,130,50,8
     EDITTEXT        IDC_INTERNAL_NAME,61,128,90,12,ES_AUTOHSCROLL | ES_READONLY | NOT WS_TABSTOP
     DEFPUSHBUTTON   "OK",IDOK,251,126,50,14
     PUSHBUTTON      "Annuler",IDCANCEL,312,126,50,14
+    CONTROL         "Exporte BD Normals",IDC_EXPORT_NORMALS,"Button",BS_AUTOCHECKBOX | WS_TABSTOP,13,68,79,10
+    EDITTEXT        IDC_DB_NAME,53,85,296,14,ES_AUTOHSCROLL
 END
 
 IDD_SIM_LOC DIALOGEX 0, 0, 257, 228
@@ -6111,6 +6113,8 @@ BEGIN
     CONTROL         "Export match stations",IDC_EXPORT_MATCH,"Button",BS_AUTOCHECKBOX | WS_TABSTOP,13,68,83,10
     LTEXT           "Name:",IDC_CMN_STATIC2,18,88,22,8
     EDITTEXT        IDC_MATCH_NAME,53,85,296,14,ES_AUTOHSCROLL
+    CONTROL         "Exporte Normals' DB",IDC_EXPORT_NORMALS,"Button",BS_AUTOCHECKBOX | WS_TABSTOP,13,68,81,10
+    EDITTEXT        IDC_DB_NAME,53,85,296,14,ES_AUTOHSCROLL
 END
 
 IDD_SIM_LOC DIALOGEX 0, 0, 257, 228
@@ -6626,9 +6630,9 @@ BEGIN
     CONTROL         "Fix Fo",IDC_FIT_FIXE_F0,"Button",BS_AUTOCHECKBOX | BS_MULTILINE | WS_TABSTOP,242,375,40,14
     EDITTEXT        IDC_FIT_F0_FROM,319,375,40,12,ES_AUTOHSCROLL
     CONTROL         "Use dead",IDC_FIT_USE_DEAD,"Button",BS_AUTOCHECKBOX | BS_MULTILINE | WS_TABSTOP,375,375,75,14
-    LTEXT           "Adult name",IDC_STATIC,242,318,36,8
-    EDITTEXT        IDC_FIT_ADULT_NAME,320,315,40,12,ES_AUTOHSCROLL
-    LTEXT           "Adults don't have T (hi)",IDC_STATIC,366,317,84,13
+    LTEXT           "Adult name",IDC_STATIC,242,314,36,8
+    EDITTEXT        IDC_FIT_ADULT_NAME,320,311,130,12,ES_AUTOHSCROLL
+    LTEXT           "Adults don't have T (hi)",IDC_STATIC,259,324,84,13
     COMBOBOX        IDC_FIT_SA_PRESET,302,168,119,158,CBS_DROPDOWN | WS_VSCROLL | WS_TABSTOP
     COMBOBOX        IDC_FIT_METHOD,330,188,119,158,CBS_DROPDOWN | WS_VSCROLL | WS_TABSTOP
     LTEXT           "Accuracy/time",IDC_STATIC,242,170,54,9
