+++ wbsModels/AllModels.sln
@@ -184,6 +184,8 @@ Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Leucotaraxis spp", "Leucota
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "StdPrcpETIndex", "StdPrcpETIndex\SPEIModel.vcxproj", "{E0F91835-E883-4E95-8E6E-229976026583}"
 EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CreateBioSIMDatabase", "CreateBioSIMDatabase\MakeBioSIMNormalsModel.vcxproj", "{65D980BC-E6A9-4B2E-9B99-C4DA6BA20DCF}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Win32 = Debug|Win32
@@ -920,6 +922,12 @@ Global
 		{E0F91835-E883-4E95-8E6E-229976026583}.Release|Win32.Build.0 = Release|Win32
 		{E0F91835-E883-4E95-8E6E-229976026583}.Release|x64.ActiveCfg = Release|x64
 		{E0F91835-E883-4E95-8E6E-229976026583}.Release|x64.Build.0 = Release|x64
+		{65D980BC-E6A9-4B2E-9B99-C4DA6BA20DCF}.Debug|Win32.ActiveCfg = Debug|x64
+		{65D980BC-E6A9-4B2E-9B99-C4DA6BA20DCF}.Debug|x64.ActiveCfg = Debug|x64
+		{65D980BC-E6A9-4B2E-9B99-C4DA6BA20DCF}.Debug|x64.Build.0 = Debug|x64
+		{65D980BC-E6A9-4B2E-9B99-C4DA6BA20DCF}.Release|Win32.ActiveCfg = Release|x64
+		{65D980BC-E6A9-4B2E-9B99-C4DA6BA20DCF}.Release|x64.ActiveCfg = Release|x64
+		{65D980BC-E6A9-4B2E-9B99-C4DA6BA20DCF}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
