+++ wbsModels/Climatic/CCBioModel.vcxproj.user
@@ -9,11 +9,11 @@
     <DebuggerFlavor>WindowsLocalDebugger</DebuggerFlavor>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
-    <LocalDebuggerCommand>..\..\bin\Releasex$(PlatformArchitecture)\BioSIM11.exe</LocalDebuggerCommand>
+    <LocalDebuggerCommand>..\..\bin\Releasex$(PlatformArchitecture)\bin\BioSIM11.exe</LocalDebuggerCommand>
     <DebuggerFlavor>WindowsLocalDebugger</DebuggerFlavor>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
-    <LocalDebuggerCommand>..\..\bin\Releasex$(PlatformArchitecture)\BioSIM11.exe</LocalDebuggerCommand>
+    <LocalDebuggerCommand>..\..\bin\Releasex$(PlatformArchitecture)\bin\BioSIM11.exe</LocalDebuggerCommand>
     <DebuggerFlavor>WindowsLocalDebugger</DebuggerFlavor>
   </PropertyGroup>
 </Project>
