--- wbsModels/CreateBioSIMDatabase/stdafx.cpp
@@ -1,8 +1,0 @@
-
-// stdafx.cpp : source file that includes just the standard includes
-// BioSIM.pch will be the pre-compiled header
-// stdafx.obj will contain the pre-compiled type information
- 
-#include "stdafx.h"
-
-
